refactor: rename subdomain-dao folders into sdao
