+++ package.json
@@ -4,12 +4,12 @@
   "version": "0.0.1-semantic-release",
   "scripts": {
     "compile": "SKIP_LOAD=true hardhat compile",
-    "test:sdao-erc721-generator": "hardhat test ./test/subdomain-dao/sdao-erc721-generator.spec.ts",
-    "test:sdao-erc1155-generator": "hardhat test ./test/subdomain-dao/sdao-erc1155-generator.spec.ts",
-    "test:sdao-code-accessible": "hardhat test ./test/subdomain-dao/sdao-code-accessible.spec.ts",
-    "test:sdao-claimable": "hardhat test ./test/subdomain-dao/sdao-claimable.spec.ts",
-    "test:sdao-registry": "hardhat test ./test/subdomain-dao/sdao-registry.spec.ts",
-    "test:sdao-reserved-limited": "hardhat test ./test/subdomain-dao/sdao-reserved-limited.spec.ts",
+    "test:sdao-erc721-generator": "hardhat test ./test/sdao/sdao-erc721-generator.spec.ts",
+    "test:sdao-erc1155-generator": "hardhat test ./test/sdao/sdao-erc1155-generator.spec.ts",
+    "test:sdao-code-accessible": "hardhat test ./test/sdao/sdao-code-accessible.spec.ts",
+    "test:sdao-claimable": "hardhat test ./test/sdao/sdao-claimable.spec.ts",
+    "test:sdao-registry": "hardhat test ./test/sdao/sdao-registry.spec.ts",
+    "test:sdao-reserved-limited": "hardhat test ./test/sdao/sdao-reserved-limited.spec.ts",
     "test:ens-label-booker": "hardhat test ./test/ens-label-booker.spec.ts",
     "test:ens-registry": "hardhat test ./test/ens-registry.spec.ts",
     "test": "hardhat test ./test/*.spec.ts ./test/**/*.spec.ts",
