+++ tasks/deploy/index.ts
@@ -1,7 +1,7 @@
 export * from './deploy-ens-full';
 export * from './deploy-label-booker';
-export * from './subdomain-dao/deploy-sdao-preset-claimable';
-export * from './subdomain-dao/deploy-sdao-preset-code-accessible';
-export * from './subdomain-dao/deploy-sdao-preset-erc721';
-export * from './subdomain-dao/deploy-sdao-preset-erc1155';
-export * from './subdomain-dao/deploy-sdao';
+export * from './sdao/deploy-sdao-preset-claimable';
+export * from './sdao/deploy-sdao-preset-code-accessible';
+export * from './sdao/deploy-sdao-preset-erc721';
+export * from './sdao/deploy-sdao-preset-erc1155';
+export * from './sdao/deploy-sdao';
